Crate main_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

AssetDeposit

AssetsStringLimit

AuctionCost

AuctionDuration

BaseFilter

BasicDeposit

BlockHashCount

BlockReward

Burn

CooloffPeriod

DataDepositPerByte

DealWithFees

DepositBase

DepositFactor

EnactmentPeriod

ExistentialDeposit

FastTrackVotingPeriod

FieldDeposit

InstantAllowed

LaunchPeriod

MaxAdditionalFields

MaxApprovals

MaxCollectivesProposalWeight

MaxLocks

MaxProposals

MaxRegistrars

MaxReserves

MaxScheduledPerBlock

MaxSignatories

MaxSubAccounts

MaxVotes

MaximumMessageSize

MaximumMultiplier

MaximumSchedulerWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MinVestedTransfer

MinimalBid

MinimumDeposit

MinimumMultiplier

MinimumPeriod

NoPreimagePostponement

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMinimum

ReferenceCallWeight

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SpendPeriod

SubAccountDeposit

TargetBlockFullness

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TransactionByteFee

TreasuryPalletId

UnvestedFundsAllowedWithdrawReasons

Version

VotingPeriod

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

WindowSize

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Constants

VERSION

Standalone runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

AssetsConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

Datalog

Democracy

DemocracyConfig

DigitalTwin

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

Header

Block header type as expected by this runtime.

Identity

Launch

Liability

Lighthouse

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

Preimage

RWS

Scheduler

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig